Adding standard gitignore, remove code duplication
